+++ packages/block-editor/src/store/reducer.js
@@ -2172,6 +2172,30 @@ function listViewExpandRevision( state = 0, action ) {
 	return state;
 }
 
+/**
+ * Reducer tracking whether the list view content panel popover is open.
+ *
+ * @param {boolean} state  Current state.
+ * @param {Object}  action Dispatched action.
+ *
+ * @return {boolean} Updated state.
+ */
+export function listViewContentPanelOpen( state = false, action ) {
+	switch ( action.type ) {
+		case 'OPEN_LIST_VIEW_CONTENT_PANEL':
+			return true;
+
+		case 'CLOSE_LIST_VIEW_CONTENT_PANEL':
+			return false;
+
+		// Close when selection is cleared
+		case 'CLEAR_SELECTED_BLOCK':
+			return false;
+	}
+
+	return state;
+}
+
 const combinedReducers = combineReducers( {
 	blocks,
 	isDragging,
@@ -2205,6 +2229,7 @@ const combinedReducers = combineReducers( {
 	hasBlockSpotlight,
 	openedListViewPanels,
 	listViewExpandRevision,
+	listViewContentPanelOpen,
 } );
 
 /**